requirements.txt
discord.py==2.4.0
python-dotenv>=1.0
loguru>=0.7
pydantic>=2.8
openai>=1.40
httpx>=0.27
PyYAML>=6.0
